remove data seeder, restructure product package, update port and environment handling, and refine metadata for better maintainability and consistency
