+++ docker-compose.yml
@@ -8,7 +8,7 @@ services:
       - "5432:5432"
     environment:
       POSTGRES_USER: postgres
-      POSTGRES_PASSWORD: (!2KvQSnQzWg)
+      POSTGRES_PASSWORD: "(!2KvQSnQzWg)"
       POSTGRES_DB: customer_service_app
     healthcheck:
       test: ["CMD-SHELL", "pg_isready -U postgres -d customer_service_app"]
