+++ src/main/resources/application.properties
@@ -1,19 +1,9 @@
 spring.application.name=customer-service-app
 
-# Server
-# Use 8090 to avoid conflicts with IDEs or other local services
-server.port=8090
-
-# Expose useful actuator endpoints for debugging (mappings helps diagnose 404s)
-management.endpoints.web.exposure.include=mappings,health,info
-
-# MVC path matching: also match trailing slashes (e.g., /products and /products/)
-spring.mvc.trailing-slash.match=true
-
 # Datasource config via env with safe defaults
-spring.datasource.url=${JDBC_URL:jdbc:postgresql://localhost:5432/customer_service_app}
-spring.datasource.username=${DB_USERNAME:postgres}
-spring.datasource.password=${DB_PASSWORD:postgres}
+spring.datasource.url=${SPRING_DATASOURCE_URL:${JDBC_URL:${DATABASE_URL:jdbc:postgresql://localhost:5432/customer_service_app}}}
+spring.datasource.username=${SPRING_DATASOURCE_USERNAME:${DB_USERNAME:${POSTGRES_USER:postgres}}}
+spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:${DB_PASSWORD:${POSTGRES_PASSWORD:(!2KvQSnQzWg)}}}
 spring.datasource.hikari.maximum-pool-size=5
 spring.sql.init.mode=never
 
@@ -37,5 +27,10 @@ logging.level.org.springframework.jdbc.core=INFO
 logging.level.org.springframework.jdbc.core.simple.JdbcClient=DEBUG
 
 # Java 21 virtual threads (Spring Boot 3.2+)
-# When enabled, Spring will use virtual threads for request handling and task executors where applicable.
 spring.threads.virtual.enabled=true
+
+# Bind server explicitly; default to 8082, overridable via SERVER_PORT env var
+server.port=${SERVER_PORT:8082}
+
+# Expose selected actuator endpoints for easier debugging via http/requests.http
+management.endpoints.web.exposure.include=health,info,mappings
