+++ src/main/resources/templates/index.html
@@ -3,10 +3,10 @@
 <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>FAMME - Premium Athletic Wear</title>
-    <meta name="description" content="Discover premium athletic wear and activewear designed for performance and style. Shop leggings, sports bras, and more." />
-    <meta name="author" content="FAMME" />
-    <meta name="keywords" content="athletic wear, activewear, leggings, sports bras, yoga pants, fitness clothing" />
+    <title>Customer Service App • Kotlin, Spring Boot, HTMX</title>
+    <meta name="description" content="Customer Service App: manage and search products with a modern Kotlin + Spring Boot + HTMX UI backed by PostgreSQL." />
+    <meta name="author" content="Customer Service App Team" />
+    <meta name="keywords" content="customer service app, products, spring boot, kotlin, htmx, postgres, thymeleaf" />
 
     <!-- Web Awesome v3 beta: single stylesheet + loader (includes tokens & icons) -->
     <link rel="stylesheet" href="https://early.webawesome.com/webawesome@3.0.0-beta.6/dist/styles/webawesome.css" />
@@ -14,8 +14,8 @@
     <!-- HTMX for partial updates -->
     <script src="https://unpkg.com/htmx.org@1.9.12"></script>
 
-    <meta property="og:title" content="FAMME - Premium Athletic Wear" />
-    <meta property="og:description" content="Discover premium athletic wear and activewear designed for performance and style." />
+    <meta property="og:title" content="Customer Service App" />
+    <meta property="og:description" content="Manage and search products with a fast HTMX + Thymeleaf UI. Powered by Kotlin, Spring Boot, and PostgreSQL." />
     <meta property="og:type" content="website" />
 
     <link rel="stylesheet" href="/css/app.css?v=3">
