+++ src/main/resources/templates/product-edit.html
@@ -8,7 +8,7 @@
   <!-- Web Awesome -->
   <link rel="stylesheet" href="https://early.webawesome.com/webawesome@3.0.0-beta.6/dist/styles/webawesome.css" />
   <script type="module" src="https://early.webawesome.com/webawesome@3.0.0-beta.6/dist/webawesome.loader.js"></script>
-  <!-- HTMX (not strictly needed here but kept for consistency) -->
+
   <script src="https://unpkg.com/htmx.org@1.9.12"></script>
 
   <link rel="stylesheet" href="/css/app.css?v=2">
